Cleanly split the code between Application and Pages
fixes #645

* Application is the main app itself (made with semantic-ui)
* Pages is the publicly visible pages and contact form (made with template-datagouv)
* Controllers for the main app inherit ApplicationController
* Controllers for the pages inherit PagesController
* Both inherit SharedController
* This lets us put common partials in views/application, views/pages, or views/shared.
