+++ app/assets/javascripts/pages/stats_charts.js
