+++ app/assets/stylesheets/application.sass
@@ -12,4 +12,5 @@
  *
  */
 
-@import "./*"
+@import application/*
+@import shared/*
