+++ app/assets/stylesheets/application/diagnoses.sass
