+++ app/assets/stylesheets/pages.sass
@@ -1,6 +1,6 @@
 @import template-data-gouv-main
 
-@import ../environment_ribbon
+@import shared/*
 
 // template overrides
 .nav__item a#experts_login
