+++ app/assets/stylesheets/shared/environment_ribbon.sass
