+++ app/assets/stylesheets/shared/user_impersonate.sass
