+++ app/helpers/person_helper.rb
@@ -2,7 +2,7 @@ module PersonHelper
   def person_block_raw(params)
     defaults = { name: nil, prefix: nil, suffix: nil, role: nil, email: nil, phone_number: nil }
     params = defaults.merge!(params)
-    render partial: 'person/person', locals: params
+    render 'person', params
   end
 
   def person_block(person, extra_params = {})
