+++ app/views/application/_flashes.html.haml
