+++ app/views/application/_footer.html.haml
