+++ app/views/application/_head.html.haml
@@ -6,7 +6,7 @@
   = javascript_include_tag 'application', data: { turbolinks_track: 'reload' }
   = javascript_include_tag 'https://browser.sentry-cdn.com/4.6.4/bundle.min.js', crossorigin: 'anonymous'
 
-  = render 'application/favicon'
+  = render 'favicon'
   = yield :head
   = sentry_script
   = matomo_script
