+++ app/views/application/_navbar.html.haml
