+++ app/views/application/_person.haml
