+++ app/views/layouts/application.html.haml
@@ -1,17 +1,17 @@
 !!!
 %html
-  = render partial: 'layouts/head'
+  = render 'head'
 
   %body
-    = render partial: 'layouts/navbar'
-    = render partial: 'layouts/environment_ribbon'
+    = render 'navbar'
+    = render 'environment_ribbon'
 
     - if current_staff_user
-      = render 'user_impersonate/header'
+      = render 'user_impersonate'
 
     %main
       .ui.main.text.container
-        = render partial: 'layouts/flashes'
+        = render 'flashes'
         = yield
 
-    = render partial: 'layouts/footer'
+    = render 'footer'
