+++ app/views/pages/_flashes.haml
