+++ app/views/pages/_footer.haml
