+++ app/views/pages/_navbar.haml
