+++ app/views/shared/_environment_ribbon.html.haml
