+++ app/views/shared/_favicon.html.erb
