+++ app/views/shared/flashes.js.haml
