+++ spec/features/application_controller_feature_spec.rb
@@ -37,30 +37,4 @@ describe 'ApplicationController specific features', type: :feature do
 
     it('redirects to diagnoses page') { expect(current_url).to eq diagnoses_url }
   end
-
-  describe 'render_error' do
-    login_user
-
-    before do
-      ENV['TEST_ERROR_RENDERING'] = 'true'
-      allow_any_instance_of(User).to receive(:sent_diagnoses).and_raise(raised_error)
-      visit diagnoses_path
-    end
-
-    after do
-      ENV['TEST_ERROR_RENDERING'] = 'false'
-    end
-
-    describe '404 error' do
-      let(:raised_error) { ActiveRecord::RecordNotFound }
-
-      it { expect(page.html).to include('Cette page n’existe pas, ou vous n’y avez pas accès.') }
-    end
-
-    describe '500 error' do
-      let(:raised_error) { ArgumentError }
-
-      it { expect(page.html).to include 'Cette erreur était inattendue…' }
-    end
-  end
 end
